Rajout interface de test basique + création de quelques controller + modif modèle

+++ includes/modele/equipe.class.php
@@ -1,14 +1,13 @@
 <?php
 require("database.class.php");
-
-
+require("../controller/equipe.controller.class.php");
 class Equipe {
 	private $_id;
 	private $_nom;
 	private $_score;
 	private $_deleted;
-	function __construct($id, $nom, $score){
-		$this->_id=$id;
+	function __construct($nom, $score){
+		$this->_id=NULL;
 		$this->_nom=$nom;
 		$this->_score=$score;
 		$this->_deleted=false;
@@ -23,19 +22,24 @@ class Equipe {
 		$this->_deleted=false;
 	}
 	function saveToDb(){
-		$database = new Database();
-		if ($_deleted)
-		{
-			$database->delete('equipe', array("id" => $this->_id));
-		}	
-		else if ($database->count('equipe', array("id" => $this->_id))) // Existe en db, on update
-		{
-			$database->update('equipe', array("id" => $this->_id), array("nom" => $this->_nom, "score" => $this->_score));
+		$controller=new Controller_Equipe($this);
+		if ($controller->isGood()){
+			$database = new Database();
+			if ($_deleted)
+			{
+				$database->delete('equipe', array("id" => $this->_id));
+			}	
+			else if ($database->count('equipe', array("id" => $this->_id))) // Existe en db, on update
+			{
+				$database->update('equipe', array("id" => $this->_id), array("nom" => $this->_nom, "score" => $this->_score));
+			}
+			else
+			{
+				$database->create('equipe', array("id" => $this->_id, "nom" => $this->_nom, "score" => $this->_score));
+			}
+			return true;
 		}
-		else
-		{
-			$database->create('equipe', array("id" => $this->_id, "nom" => $this->_nom, "score" => $this->_score));
-		}
+		return false;
 	}
 	function getId() {
 		return $this->_id;
