+++ includes/modele/partenaire.class.php
@@ -1,5 +1,6 @@
 <?php
 require("database.class.php");
+require("../controller/partenaire.controller.class.php");
 //Fonctions de la classe Partenaire 
 
 class Partenaire{
@@ -33,19 +34,24 @@ class Partenaire{
 		$this->_deleted=false;
 	}
 	function saveToDb(){
-		$database = new Database();
-		if ($_deleted)
-		{
-			$database->delete('partenaire', array("id" => $this->_id));
-		}	
-		else if ($database->count('partenaire', array("id" => $this->_id))) // Existe en db, on update
-		{
-			$database->update('partenaire', array("id" => $this->_id), array("nom" => $this->_nom, "description" => $this->_description, "mail" => $this->_mail, "url" => $this->_siteWeb, "telephone" => $this->_telephone));
+		$controller=new Controller_Equipe($this);
+		if ($controller->isGood()){
+			$database = new Database();
+			if ($_deleted)
+			{
+				$database->delete('partenaire', array("id" => $this->_id));
+			}	
+			else if ($database->count('partenaire', array("id" => $this->_id))) // Existe en db, on update
+			{
+				$database->update('partenaire', array("id" => $this->_id), array("nom" => $this->_nom, "description" => $this->_description, "mail" => $this->_mail, "url" => $this->_siteWeb, "telephone" => $this->_telephone));
+			}
+			else
+			{
+				$database->create('partenaire', array("id" => $this->_id, "nom" => $this->_nom, "description" => $this->_description, "mail" => $this->_mail, "url" => $this->_siteWeb, "telephone" => $this->_telephone));
+			}
+			return true;
 		}
-		else
-		{
-			$database->create('partenaire', array("id" => $this->_id, "nom" => $this->_nom, "description" => $this->_description, "mail" => $this->_mail, "url" => $this->_siteWeb, "telephone" => $this->_telephone));
-		}
+		return false;
 	}
 	function getId() {
 	   return $this->_id;
